+++ lib/Crypto/SelfTest/Cipher/test_pkcs1_15.py
@@ -22,12 +22,10 @@
 
 from __future__ import print_function
 
-import json
 import unittest
-from binascii import unhexlify
 
 from Crypto.PublicKey import RSA
-from Crypto.SelfTest.st_common import list_test_cases, a2b_hex, b2a_hex
+from Crypto.SelfTest.st_common import list_test_cases, a2b_hex
 from Crypto import Random
 from Crypto.Cipher import PKCS1_v1_5 as PKCS
 from Crypto.Util.py3compat import b
@@ -38,15 +36,17 @@ from Crypto.SelfTest.loader import load_test_vectors_wycheproof
 def rws(t):
     """Remove white spaces, tabs, and new lines from a string"""
     for c in ['\n', '\t', ' ']:
-        t = t.replace(c,'')
+        t = t.replace(c, '')
     return t
+
 
 def t2b(t):
     """Convert a text string with bytes in hex form to a byte string"""
     clean = b(rws(t))
-    if len(clean)%2 == 1:
+    if len(clean) % 2 == 1:
         raise ValueError("Even number of characters expected")
     return a2b_hex(clean)
+
 
 class PKCS1_15_Tests(unittest.TestCase):
 
@@ -131,51 +131,72 @@ HKukWBcq9f/UOmS0oEhai/6g+Uf7VHJdWaeO5LzuvwU=
                 self.assertRaises(ValueError, cipher.encrypt, pt)
 
         def testVerify1(self):
-                for test in self._testData:
-                        # Build the key
-                        key = RSA.importKey(test[0])
-                        # The real test
-                        cipher = PKCS.new(key)
-                        pt = cipher.decrypt(t2b(test[2]), "---")
-                        self.assertEqual(pt, b(test[1]))
+            for test in self._testData:
+                key = RSA.importKey(test[0])
+                expected_pt = b(test[1])
+                ct = t2b(test[2])
+                cipher = PKCS.new(key)
+
+                # The real test
+                pt = cipher.decrypt(ct, None)
+                self.assertEqual(pt, expected_pt)
+
+                pt = cipher.decrypt(ct, b'\xFF' * len(expected_pt))
+                self.assertEqual(pt, expected_pt)
 
         def testVerify2(self):
-                # Verify that decryption fails if ciphertext is not as long as
-                # RSA modulus
-                cipher = PKCS.new(self.key1024)
-                self.assertRaises(ValueError, cipher.decrypt, '\x00'*127, "---")
-                self.assertRaises(ValueError, cipher.decrypt, '\x00'*129, "---")
+            # Verify that decryption fails if ciphertext is not as long as
+            # RSA modulus
+            cipher = PKCS.new(self.key1024)
+            self.assertRaises(ValueError, cipher.decrypt, '\x00'*127, "---")
+            self.assertRaises(ValueError, cipher.decrypt, '\x00'*129, "---")
 
-                # Verify that decryption fails if there are less then 8 non-zero padding
-                # bytes
-                pt = b('\x00\x02' + '\xFF'*7 + '\x00' + '\x45'*118)
-                pt_int = bytes_to_long(pt)
-                ct_int = self.key1024._encrypt(pt_int)
-                ct = long_to_bytes(ct_int, 128)
-                self.assertEqual("---", cipher.decrypt(ct, "---"))
+            # Verify that decryption fails if there are less then 8 non-zero padding
+            # bytes
+            pt = b('\x00\x02' + '\xFF'*7 + '\x00' + '\x45'*118)
+            pt_int = bytes_to_long(pt)
+            ct_int = self.key1024._encrypt(pt_int)
+            ct = long_to_bytes(ct_int, 128)
+            self.assertEqual(b"---", cipher.decrypt(ct, b"---"))
 
         def testEncryptVerify1(self):
-                # Encrypt/Verify messages of length [0..RSAlen-11]
-                # and therefore padding [8..117]
-                for pt_len in range(0,128-11+1):
-                    pt = self.rng(pt_len)
-                    cipher = PKCS.new(self.key1024)
-                    ct = cipher.encrypt(pt)
-                    pt2 = cipher.decrypt(ct, "---")
-                    self.assertEqual(pt,pt2)
+            # Encrypt/Verify messages of length [0..RSAlen-11]
+            # and therefore padding [8..117]
+            for pt_len in range(0, 128 - 11 + 1):
+                pt = self.rng(pt_len)
+                cipher = PKCS.new(self.key1024)
+                ct = cipher.encrypt(pt)
+                pt2 = cipher.decrypt(ct, b'\xAA' * pt_len)
+                self.assertEqual(pt, pt2)
+
+        def test_encrypt_verify_exp_pt_len(self):
+
+            cipher = PKCS.new(self.key1024)
+            pt = b'5' * 16
+            ct = cipher.encrypt(pt)
+            sentinel = b'\xAA' * 16
+
+            pt_A = cipher.decrypt(ct, sentinel, 16)
+            self.assertEqual(pt, pt_A)
+
+            pt_B = cipher.decrypt(ct, sentinel, 15)
+            self.assertEqual(sentinel, pt_B)
+
+            pt_C = cipher.decrypt(ct, sentinel, 17)
+            self.assertEqual(sentinel, pt_C)
 
         def testByteArray(self):
             pt = b"XER"
             cipher = PKCS.new(self.key1024)
             ct = cipher.encrypt(bytearray(pt))
-            pt2 = cipher.decrypt(bytearray(ct), "---")
+            pt2 = cipher.decrypt(bytearray(ct), '\xFF' * len(pt))
             self.assertEqual(pt, pt2)
 
         def testMemoryview(self):
             pt = b"XER"
             cipher = PKCS.new(self.key1024)
             ct = cipher.encrypt(memoryview(bytearray(pt)))
-            pt2 = cipher.decrypt(memoryview(bytearray(ct)), "---")
+            pt2 = cipher.decrypt(memoryview(bytearray(ct)), b'\xFF' * len(pt))
             self.assertEqual(pt, pt2)
 
 
@@ -216,14 +237,14 @@ class TestVectorsWycheproof(unittest.TestCase):
 
     def test_decrypt(self, tv):
         self._id = "Wycheproof Decrypt PKCS#1v1.5 Test #%s" % tv.id
-
+        sentinel = b'\xAA' * max(3, len(tv.msg))
         cipher = PKCS.new(tv.rsa_key)
         try:
-            pt = cipher.decrypt(tv.ct, sentinel=b'---')
+            pt = cipher.decrypt(tv.ct, sentinel=sentinel)
         except ValueError:
             assert not tv.valid
         else:
-            if pt == b'---':
+            if pt == sentinel:
                 assert not tv.valid
             else:
                 assert tv.valid
@@ -245,8 +266,10 @@ def get_tests(config={}):
     tests += [TestVectorsWycheproof(wycheproof_warnings, skip_slow_tests)]
     return tests
 
+
 if __name__ == '__main__':
-    suite = lambda: unittest.TestSuite(get_tests())
+    def suite():
+        return unittest.TestSuite(get_tests())
     unittest.main(defaultTest='suite')
 
 # vim:set ts=4 sw=4 sts=4 expandtab: