+++ lib/Crypto/SelfTest/Util/test_number.py
@@ -24,9 +24,15 @@
 
 """Self-tests for (some of) Crypto.Util.number"""
 
-from Crypto.Util.py3compat import *
-
+import math
 import unittest
+
+from Crypto.Util.py3compat import *
+from Crypto.SelfTest.st_common import list_test_cases
+
+from Crypto.Util import number
+from Crypto.Util.number import long_to_bytes
+
 
 class MyError(Exception):
     """Dummy exception used for tests"""
@@ -35,10 +41,6 @@ class MyError(Exception):
 # if any inputs cause a test failure, we'll be able to tell which ones.
 
 class MiscTests(unittest.TestCase):
-    def setUp(self):
-        global number, math
-        from Crypto.Util import number
-        import math
 
     def test_ceil_div(self):
         """Util.number.ceil_div"""
@@ -132,9 +134,35 @@ class MiscTests(unittest.TestCase):
         self.assertRaises(ValueError, number.size, -1)
 
 
+class LongTests(unittest.TestCase):
+
+    def test1(self):
+        self.assertEqual(long_to_bytes(0), b'\x00')
+        self.assertEqual(long_to_bytes(1), b'\x01')
+        self.assertEqual(long_to_bytes(0x100), b'\x01\x00')
+        self.assertEqual(long_to_bytes(0xFF00000000), b'\xFF\x00\x00\x00\x00')
+        self.assertEqual(long_to_bytes(0xFF00000000), b'\xFF\x00\x00\x00\x00')
+        self.assertEqual(long_to_bytes(0x1122334455667788), b'\x11\x22\x33\x44\x55\x66\x77\x88')
+        self.assertEqual(long_to_bytes(0x112233445566778899), b'\x11\x22\x33\x44\x55\x66\x77\x88\x99')
+
+    def test2(self):
+        self.assertEqual(long_to_bytes(0, 1), b'\x00')
+        self.assertEqual(long_to_bytes(0, 2), b'\x00\x00')
+        self.assertEqual(long_to_bytes(1, 3), b'\x00\x00\x01')
+        self.assertEqual(long_to_bytes(0x100, 1), b'\x01\x00')
+        self.assertEqual(long_to_bytes(0xFF00000001, 6), b'\x00\xFF\x00\x00\x00\x01')
+        self.assertEqual(long_to_bytes(0xFF00000001, 8), b'\x00\x00\x00\xFF\x00\x00\x00\x01')
+        self.assertEqual(long_to_bytes(0xFF00000001, 10), b'\x00\x00\x00\x00\x00\xFF\x00\x00\x00\x01')
+        self.assertEqual(long_to_bytes(0xFF00000001, 11), b'\x00\x00\x00\x00\x00\x00\xFF\x00\x00\x00\x01')
+
+    def test_err1(self):
+        self.assertRaises(ValueError, long_to_bytes, -1)
+
+
 def get_tests(config={}):
-    from Crypto.SelfTest.st_common import list_test_cases
-    tests = list_test_cases(MiscTests)
+    tests = []
+    tests += list_test_cases(MiscTests)
+    tests += list_test_cases(LongTests)
     return tests
 
 if __name__ == '__main__':